Move `split_` methods to SqlAlchemyDataSplitter and remove unused table_name param
